Add support for PP
Because the inspect method was overridden, Ruby's PP module (and by
extension pry's pretty printer) would not pretty print any objects
properly, instead leaving them on a single line which is hard to use.
This commit allows PP to work properly.

By happenstance it also greatly simplifies the CustomInspectMixin by
using the built-in pretty_print_inspect as the custom inspect method.
The removal of instance variables was the one feature that
CustomInspectMixin provided, but it turns out PP already has support for
that, so we can just use it.