Loosen faraday to 1.x
Faraday 1.0.1 was released 3/29/2020
Faraday 1.10.3 was released 1/18/2023
(it is not in the new branch, but it is more up to date)

This resolves the following warnings

```
Top level ::CompositeIO is deprecated, require 'multipart/post' and use `Multipart::Post::CompositeReadIO` instead!
Top level ::Parts is deprecated, require 'multipart/post' and use `Multipart::Post::Parts` instead!
```